refactor: apply MVC pattern
- MVC패턴을 기반으로 파일 분리
- 타이머가 멈추는 문제를 런루프의 추가로 해결
